+++ sshd-core/src/test/java/org/apache/sshd/client/config/hosts/HostConfigEntryResolverTest.java
@@ -94,7 +94,7 @@ public class HostConfigEntryResolverTest extends BaseTestSupport {
     @Test
     public void testEffectiveHostConfigResolution() throws Exception {
         HostConfigEntry entry = new HostConfigEntry(getCurrentTestName(), TEST_LOCALHOST, port, getCurrentTestName());
-        client.setHostConfigEntryResolver((host, portValue, username, context) -> entry);
+        client.setHostConfigEntryResolver((host, portValue, lclAddress, username, context) -> entry);
         client.start();
 
         try (ClientSession session = client.connect(
@@ -171,7 +171,7 @@ public class HostConfigEntryResolverTest extends BaseTestSupport {
         String host = getClass().getSimpleName();
         HostConfigEntry entry = new HostConfigEntry(host, TEST_LOCALHOST, port, user);
         entry.addIdentity(clientIdentity);
-        client.setHostConfigEntryResolver((host1, portValue, username, context) -> entry);
+        client.setHostConfigEntryResolver((host1, portValue, lclAddress, username, context) -> entry);
 
         client.start();
         try (ClientSession session = client.connect(