+++ CHANGES.md
@@ -98,6 +98,9 @@ block size.
 
     * This method is now also implemented by the `Identity` interface as well
 
+* `HostConfigEntryResolver#resolveEffectiveHost` accepts also an (optional) initial connection
+context and/or local peer binding address - propagated from the `ClientSessionCreator#connect` invocation.
+
 ## Behavioral changes and enhancements
 
 * [SSHD-708](https://issues.apache.org/jira/browse/SSHD-708) - Add support for password encrypted OpenSSH private key files.